Added an experimental, ABI-stable, raw C interface To enable easier binding from other languages/FFI interfaces. Can be built with -Dbuild_abi=ON, linked against with -ldart_abi Include file is #include <dart/abi.h> Tests are currently minimal and documentation is yet to come.
