+++ test/CMakeLists.txt
@@ -2,23 +2,28 @@
 add_library(catch OBJECT catch_driver.cc)
 
 # Setup targets for our unit tests.
+add_executable(unit_tests type_unit_tests.cc obj_unit_tests.cc arr_unit_tests.cc str_unit_tests.cc int_unit_tests.cc dcm_unit_tests.cc bool_unit_tests.cc null_unit_tests.cc iteration_unit_tests.cc ptr_unit_tests.cc misc_unit_tests.cc optional_unit_tests.cc string_view_unit_tests.cc $<TARGET_OBJECTS:catch>)
 if (build_abi)
   add_executable(abi_tests abi_unit_tests.cc $<TARGET_OBJECTS:catch>)
   target_link_libraries(abi_tests dart_abi)
   target_compile_options(abi_tests PUBLIC "-g")
+  target_compile_options(unit_tests PUBLIC "-DDART_HAS_ABI")
 endif ()
-add_executable(unit_tests type_unit_tests.cc obj_unit_tests.cc arr_unit_tests.cc str_unit_tests.cc int_unit_tests.cc dcm_unit_tests.cc bool_unit_tests.cc null_unit_tests.cc iteration_unit_tests.cc ptr_unit_tests.cc misc_unit_tests.cc optional_unit_tests.cc string_view_unit_tests.cc $<TARGET_OBJECTS:catch>)
 
 if (can_build_yaml)
   target_link_libraries(unit_tests ${libyaml})
-  target_link_libraries(abi_tests ${libyaml})
+  if (build_abi)
+    target_link_libraries(abi_tests ${libyaml})
+  endif ()
 endif ()
 if (extended_test)
   target_compile_options(unit_tests PUBLIC "-DDART_EXTENDED_TESTS")
 endif ()
 if (32bit)
   target_link_libraries(unit_tests atomic)
-  target_link_libraries(abi_tests atomic)
+  if (build_abi)
+    target_link_libraries(abi_tests atomic)
+  endif ()
 endif ()
 if (static_test)
   target_link_libraries(unit_tests "-static-libstdc++")
@@ -34,6 +39,10 @@ if (librj)
   target_compile_options(json_test PUBLIC "-g")
   if (extended_test)
     target_compile_options(json_test PUBLIC "-DDART_EXTENDED_TESTS")
+  endif ()
+  if (build_abi)
+    target_compile_options(json_test PUBLIC "-DDART_HAS_ABI")
+    target_link_libraries(json_test dart_abi)
   endif ()
 endif ()
